Revert "move budget_record struct to db_maint"
This reverts commit 9a4e745052fa7a209a44bad4affd30923918ad6e.
